creating download function, zip files and delete function/message

+++ app/Http/Controllers/FileController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\DestroyFilesRequest;
+use App\Http\Requests\FilesActionRequest;
 use App\Http\Requests\StoreFileRequest;
 use App\Http\Requests\StoreFolderRequest;
 use App\Http\Resources\FileResource;
@@ -10,7 +10,10 @@ use App\Models\File;
 use Illuminate\Auth\Events\Validated;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Storage;
 use Inertia\Inertia;
+use Illuminate\Support\Str;
+use ZipArchive;
 
 class FileController extends Controller
 {
@@ -109,7 +112,7 @@ class FileController extends Controller
         }
     }
 
-    public function destroy(DestroyFilesRequest $request)
+    public function destroy(FilesActionRequest $request)
     {
         $data = $request->validated();
         $parent = $request->parent;
@@ -124,12 +127,91 @@ class FileController extends Controller
             foreach($data['ids'] ?? [] as $id)
             {
                 $file = File::find($id);
-                $file->delete();
+                if($file)
+                {
+                    $file->delete();
+                }
             }
         }
 
-        return to_route('myFiles', ['folder' => $parent->path]);
+        return to_route('MyFiles', ['folder' => $parent->path]);
     }
+
+    public function createZip($files): string
+    {
+        $zipPath = 'zip/'. Str::random().'zip';
+        $publicPath = "public/$zipPath";
+
+        if(!is_dir(dirname($publicPath)))
+        {
+            Storage::makeDirectory(dirname(($publicPath)));
+        }
+
+        $zipFile = Storage::path($publicPath);
+
+        $zip = new \ZipArchive();
+
+        if ($zip->open($zipFile, \ZipArchive::CREATE | \ZipArchive::OVERWRITE) === true) {
+            $this->addFilesToZip($zip, $files);
+        }
+
+        $zip->close();
+
+        return asset(Storage::url($zipPath));
+    }
+
+    public function download(FilesActionRequest $request)
+    {
+        $data = $request->validated();
+        $parent = $request->parent;
+
+        $all = $data['all'] ?? false;
+        $ids = $data['ids'] ?? [];
+
+        if(!$all && empty($ids)) {
+            return [
+                'message' => 'Please select files to download'
+            ];
+        }
+
+        if($all) {
+            $url = $this->createZip($parent->children);
+            $filename = $parent->name . '.zip';
+        } else {
+            if(count($ids) === 1){
+                $file = File::find($ids[0]);
+                if($file->is_folder){
+                    // is a folder
+                    if($file->children->count() === 0 )
+                    {
+                        return [
+                            'message' => 'The folder is empty'
+                        ];
+                    }
+                    $url = $this->createZip($file->children);
+                    $filename = $file->name . '.zip';
+                } else {
+                    // Its a file
+                    $destination = 'public/'. pathinfo($file->storage_path, PATHINFO_BASENAME);
+                    Storage::copy($file->storage_path, $destination);
+                    
+                    $url = asset(Storage::url($destination));
+                    $filename = $file->name;
+                }
+            } else {
+                $files = File::query()->whereIn('id', $ids)->get();
+
+                $url = $this->createZip($files);
+                $filename = $parent->name.'.zip';
+            }
+        }
+
+        return [
+            'url' => $url,
+            'filename' => $filename
+        ];
+    }
+
 
     private function saveFile($file, $user, $parent): void
     {
@@ -144,5 +226,17 @@ class FileController extends Controller
 
         $parent->appendNode($model);
 
+    }
+
+    private function addFilesToZip($zip, $files, $ancestors = '')
+    {
+        foreach($files as $file){
+            if($file->is_folder)
+            {
+                $this->addFilesToZip($zip, $file->children, $ancestors . $file->name . '/');
+            } else {
+                $zip->addFile(Storage::path($file->storage_path), $ancestors . $file->name);
+            }
+        }
     }
 }
